Improving Scheduler's Clock method

+++ Scheduler.h
@@ -16,6 +16,6 @@ private :
 
 public:
     Scheduler();
-    void wait(int time);
+    void wait(float time);
 };
